refactor: auction type 정리
