+++ src/components/order/OrderLostProduct.tsx
@@ -1,6 +1,7 @@
-import type { IUserAuctionLostItem } from '@/@types/AuctionItem';
 import { ParticipantCount, Price } from '@/shared';
-import ProductItem from '../../entities/product/ui/ProductItem';
+
+import type { IUserAuctionLostItem } from '@/entities';
+import ProductItem from '../product/ui/ProductItem';
 
 const OrderLostProduct = ({ product }: { product: IUserAuctionLostItem }) => {
