+++ src/features/bid/ui/BidForm.tsx
@@ -1,14 +1,14 @@
+import { MAX_BID_COUNT, getBidSchema } from "@/features/bid/config";
+import { Button, FormField, convertCurrencyToNumber, formatCurrencyWithWon } from "@/shared";
 import { SubmitHandler, useForm } from "react-hook-form";
 
 import { Layout } from "@/app/layout/index";
-import { MAX_BID_COUNT, getBidSchema } from "@/features/bid/config";
 import { useEditableNumberInput } from "@/features/register/lib/useEditableNumberInput";
-import { Button, FormField, convertCurrencyToNumber, formatCurrencyWithWon } from "@/shared";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useState } from "react";
 import { z } from "zod";
+import AuctionItem from "../../../components/auction/ui/AuctionItem";
 import { useGetAuctionDetails } from "../../../components/details/queries";
-import AuctionItem from "../../../entities/auction/ui/AuctionItem";
 import { Input } from "../../../shared/shadcn/ui/input";
 import { usePostBid } from "../model/usePostBid";
 import { BidCaution } from "./BidCaution";
