+++ src/hooks/useEndAuction.ts
@@ -1,4 +1,4 @@
-import { IAuctionDetails } from '@/@types/AuctionDetails';
+import type { IAuctionDetails } from '@/entities';
 import { QUERY_KEYS } from '@/shared/constants/queryKeys';
 import { useQueryClient } from '@tanstack/react-query';
 import { useCallback } from 'react';
