+++ src/pages/home/api/getBestAuctions.ts
@@ -1,5 +1,6 @@
-import type { IAuctionItem } from '@/@types/AuctionItem';
 import { API_END_POINT, EmptyError, httpClient } from '@/shared';
+
+import type { IAuctionItem } from '@/entities';
 
 export const getBestAuctions = async (): Promise<IAuctionItem[]> => {
   const response = await httpClient.get(`${API_END_POINT.BEST}`);
