+++ src/pages/home/api/getPreAuctions.ts
@@ -1,5 +1,6 @@
-import type { IPreAuctionItem } from '@/@types/AuctionItem';
 import { API_END_POINT, EmptyError, httpClient } from '@/shared';
+
+import type { IPreAuctionItem } from '@/entities';
 
 export const getPreAuctions = async (): Promise<IPreAuctionItem[]> => {
   const response = await httpClient.get(`${API_END_POINT.PRE_AUCTION}?sort=most-liked,product-newest&page=0&size=5`);
